+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.EmulatorTests/Utils/TestCommon.cs
@@ -43,6 +43,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
 
         private static readonly int serverStalenessIntervalInSeconds;
         private static readonly int masterStalenessIntervalInSeconds;
+        public static readonly CosmosSerializer Serializer = new CosmosJsonDotNetSerializer();
 
         static TestCommon()
         {
@@ -67,6 +68,14 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
             }
 
             return cosmosClientBuilder.Build();
+        }
+
+        internal static CosmosClient CreateCosmosClient(CosmosClientOptions clientOptions)
+        {
+            string authKey = ConfigurationManager.AppSettings["MasterKey"];
+            string endpoint = ConfigurationManager.AppSettings["GatewayEndpoint"];
+
+            return new CosmosClient(endpoint, authKey, clientOptions);
         }
 
         internal static CosmosClient CreateCosmosClient(