+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/MockCosmosUtil.cs
@@ -20,6 +20,8 @@ namespace Microsoft.Azure.Cosmos.Client.Core.Tests
 
     internal class MockCosmosUtil
     {
+        public static readonly CosmosSerializer Serializer = new CosmosJsonDotNetSerializer();
+
         public static CosmosClient CreateMockCosmosClient(
             Action<CosmosClientBuilder> customizeClientBuilder = null,
             Cosmos.ConsistencyLevel? accountConsistencyLevel = null)